+++ FINLAY_SMITH/Main.java
@@ -1,10 +1,8 @@
-
-
 public class Main {
 
     public static void main(String[] args) {
-        // smaller y size and larger x size makes game easier
-        final int x = 2000, y = 600;
+        // smaller y size makes game much easier
+        final int x = 1000, y = 600;
 
         SpaceArena arena = new SpaceArena(x, y);
         // arena.setSeed(/*YOUR SEED HERE - otherwise RANDOM SEED*/);
