+++ FINLAY_SMITH/SpaceArena.java
@@ -1,4 +1,7 @@
 
+import GameArenaClasses.GameArena;
+import GameArenaClasses.Text;
+
 import java.util.*;
 import java.util.concurrent.TimeUnit;
